@@ -13,11 +13,6 @@ public class Main2Activity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        getWindow().getDecorView().post(new Runnable() {
-            @Override
-            public void run() {
-                TLog.i("(Main2Activity.java:17) .run->" + TLog.valueOf(8000));
-            }
-        });
+        getWindow().getDecorView().post(() -> TLog.i("(Main2Activity.java:17) .run->" + TLog.valueOf(8000)));
     }
 }
