@@ -6,14 +6,12 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 
 import org.tinlone.demo.rxjavasample.R;
-import org.tinlone.demo.rxjavasample.activity.view.ViewATestActivity;
-import org.tinlone.demo.rxjavasample.activity.view.ViewBTestActivity;
-import org.tinlone.demo.rxjavasample.activity.view.ViewCTestActivity;
 import org.tinlone.demo.rxjavasample.adapter.MyAdapter;
+import org.tinlone.demo.rxjavasample.config.ActivitiesListConfig;
 
-import java.util.ArrayList;
-import java.util.List;
-
+/**
+ * @author Administrator
+ */
 public class MyActivity extends AppCompatActivity {
 
     private RecyclerView rvList;
@@ -23,20 +21,11 @@ public class MyActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_my);
         initView();
-        List<String> data = new ArrayList<>();
-        data.add("自定义控件学习1");
-        data.add("自定义控件学习2");
-        data.add("自定义控件学习3");
-
-        List<Class> activities = new ArrayList<>();
-        activities.add(ViewATestActivity.class);
-        activities.add(ViewBTestActivity.class);
-        activities.add(ViewCTestActivity.class);
         rvList.setLayoutManager(new LinearLayoutManager(this));
-        rvList.setAdapter(new MyAdapter(this,data,activities));
+        rvList.setAdapter(new MyAdapter(this, ActivitiesListConfig.getTitles(),ActivitiesListConfig.getAvtivities()));
     }
 
     private void initView() {
-        rvList = (RecyclerView) findViewById(R.id.rv_list);
+        rvList = findViewById(R.id.rv_list);
     }
 }
